Function App/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
azure-storage-blob
pandas
python-dotenv==1.0.1
langchain-openai==0.2.0